Update : 생성자 함수를 활용하기 <description> 생성자 함수를 활용하기 위해 baseball.js 파일을 만들었고, 필요한 것들을 인스턴스의 메소드를 활용하는 것으로 구현해봤다. 아무래도 생성자 함수와 export를 활용하라고 한 뜻이 이것을 의미하는 것 같아 바꾸어 봤다.
