+++ index.html
@@ -19,5 +19,6 @@
       <div id="result"></div>
     </div>
     <script type="module" src="src/index.js"></script>
+    <script type="module" src="src/baseball.js"></script>
   </body>
 </html>
